Check for low disk space from within Python
I was checking for low disk space in a Bash script, and parsing the
output of running `df`. I think this was taking a lot of time because I
was starting a bunch of processes, and may have been causing the video
to skip a little bit, so just check from in Python now between saving
videos.