Parse Direwolf output
So when you run Direwolf, it prints the received APRS messages to the
terminal. I was hoping I could just parse stdout or stderr from the
process, but it looks like Direwolf actually uses something else to
display these? So then I thought, I'll just have it log to a logfile and
then tail that file. And that sort of worked, except we don't get the
original APRS message, just a parsed version. So I had to work around
that garbage.